Chat/requirements.txt
amqp==5.2.0
anyio==4.1.0
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
billiard==4.2.0
celery==5.3.5
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==15.1.0
cron-descriptor==1.4.0
cryptography==40.0.2
daphne==4.0.0
Django==4.2.1
django-celery-beat==2.5.0
django-cors-headers==4.3.1
django-rest-knox==4.2.0
django-timezone-field==6.1.0
djangorestframework==3.14.0
drf-spectacular==0.26.2
exceptiongroup==1.2.0
h11==0.14.0
httptools==0.6.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
jsonschema==4.17.3
kombu==5.3.4
msgpack==1.0.5
prompt-toolkit==3.0.41
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.1.1
pyrsistent==0.19.3
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
redis==4.5.5
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
uvicorn==0.24.0.post1
uvloop==0.19.0
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.10
websockets==12.0
zope.interface==6.0
